+++ CHANGELOG.md
@@ -44,6 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `hover` property and selector state to all widgets
 * Add `on_enter` and `on_leave` callback to `MouseHandler`
 * Refactor theming (cleanup default theme, introduce state order)
+* Fix theming related bugs
+* Register fonts on `Theme` struct
+* Update OrbTks default themes
+* New redox theme crate
 
 ### 0.3.1-alpha3